02 계산기 구현
